Removing duplicate DocumentTools and DocumentUtils
