+++ gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/tasks/explode/loader/EntityLoaderHelper.java
@@ -19,7 +19,7 @@ import java.util.Map;
 import java.util.Objects;
 
 import static com.ca.apim.gateway.cagatewayconfig.util.gateway.BundleElementNames.*;
-import static com.ca.apim.gateway.cagatewayexport.util.xml.DocumentUtils.getChildElements;
+import static com.ca.apim.gateway.cagatewayconfig.util.xml.DocumentUtils.getChildElements;
 import static java.lang.Integer.parseInt;
 import static java.lang.Long.parseLong;
 import static java.util.Collections.emptyMap;
