+++ gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/tasks/explode/loader/IdentityProviderLoader.java
@@ -18,11 +18,11 @@ import java.util.List;
 import java.util.Map;
 
 import static com.ca.apim.gateway.cagatewayconfig.util.gateway.BundleElementNames.*;
+import static com.ca.apim.gateway.cagatewayconfig.util.xml.DocumentUtils.*;
 import static com.ca.apim.gateway.cagatewayexport.tasks.explode.bundle.entity.IdentityProviderEntity.Builder;
 import static com.ca.apim.gateway.cagatewayexport.tasks.explode.bundle.entity.IdentityProviderEntity.Type;
 import static com.ca.apim.gateway.cagatewayexport.tasks.explode.bundle.entity.IdentityProviderEntity.Type.fromType;
 import static com.ca.apim.gateway.cagatewayexport.tasks.explode.loader.EntityLoaderHelper.mapPropertiesElements;
-import static com.ca.apim.gateway.cagatewayexport.util.xml.DocumentUtils.*;
 
 @Singleton
 public class IdentityProviderLoader implements EntityLoader<IdentityProviderEntity> {
