--- gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/util/file/DocumentFileUtilsException.java
@@ -1,17 +1,0 @@
-/*
- * Copyright (c) 2018 CA. All rights reserved.
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE file for details.
- */
-
-package com.ca.apim.gateway.cagatewayexport.util.file;
-
-public class DocumentFileUtilsException extends RuntimeException {
-    public DocumentFileUtilsException(String message) {
-        super(message);
-    }
-
-    public DocumentFileUtilsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
