--- gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/util/xml/DocumentParseException.java
@@ -1,13 +1,0 @@
-/*
- * Copyright (c) 2018 CA. All rights reserved.
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE file for details.
- */
-
-package com.ca.apim.gateway.cagatewayexport.util.xml;
-
-public class DocumentParseException extends Exception {
-    public DocumentParseException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
