--- gateway-export-plugin/src/main/java/com/ca/apim/gateway/cagatewayexport/util/xml/DocumentToolsException.java
@@ -1,18 +1,0 @@
-/*
- * Copyright (c) 2018 CA. All rights reserved.
- * This software may be modified and distributed under the terms
- * of the MIT license.  See the LICENSE file for details.
- */
-
-package com.ca.apim.gateway.cagatewayexport.util.xml;
-
-public class DocumentToolsException extends RuntimeException {
-
-    public DocumentToolsException(String message) {
-        super(message);
-    }
-
-    public DocumentToolsException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
